fix(tuneshift): resolve ruff B/RUF correctness lint findings
Address the behavior-sensitive lint residue after the format pass:

- B905: add strict=True to the four co-indexed zip() calls (column/row and
  scope/layer pairs are equal-length by construction), so a future length
  mismatch fails loudly instead of silently truncating.
- B023: bind the loop `offset` via a default arg in the two Spotify paging
  lambdas, matching the existing `lambda batch=batch:` convention and removing
  the late-binding closure hazard.
- B904: raise SystemExit(1) from None on invalid interactive selection so the
  ValueError/IndexError is not chained as an unexpected error.
- B007: prefix unused loop control variables with underscore.
- RUF059: prefix unused unpacked variables with underscore.
- RUF005: use iterable unpacking instead of list concatenation.
- RUF046: drop redundant int() around round() (round() already returns int).

All changes are behavior-preserving; affected module tests pass (541).

+++ tools/tuneshift/tuneshift/matching/album.py
@@ -53,7 +53,7 @@ ALBUM_THRESHOLDS = RecommendationThresholds(
 
 def _signed_points(penalty: float, weight: int) -> int:
     """A readable signed contribution for explainability/breakdown."""
-    return -int(round(penalty * weight))
+    return -round(penalty * weight)
 
 
 def _title_signal(source_album: str, candidate_album: str) -> SignalPenalty:
